+++ src/PowerSimulations.jl
@@ -441,7 +441,6 @@ include("core/operation_model_abstract_types.jl")
 include("core/optimization_container_types.jl")
 include("core/abstract_feedforward.jl")
 include("core/optimization_container_keys.jl")
-include("network_models/powermodels_formulations.jl")
 include("core/network_model.jl")
 include("core/parameters.jl")
 include("core/service_model.jl")
