+++ src/core/formulations.jl
@@ -16,3 +16,138 @@ struct ThermalMultiStartUnitCommitment <: AbstractCompactUnitCommitment end
 struct ThermalCompactUnitCommitment <: AbstractCompactUnitCommitment end
 struct ThermalBasicCompactUnitCommitment <: AbstractCompactUnitCommitment end
 struct ThermalCompactDispatch <: AbstractThermalDispatchFormulation end
+
+############################# Electric Load Formulations ###################################
+abstract type AbstractLoadFormulation <: AbstractDeviceFormulation end
+abstract type AbstractControllablePowerLoadFormulation <: AbstractLoadFormulation end
+struct StaticPowerLoad <: AbstractLoadFormulation end
+struct InterruptiblePowerLoad <: AbstractControllablePowerLoadFormulation end
+struct DispatchablePowerLoad <: AbstractControllablePowerLoadFormulation end
+
+############################### AC Branch Formulations #####################################
+"""
+Branch type to add unbounded flow variables and use flow constraints
+"""
+struct StaticBranch <: AbstractBranchFormulation end
+"""
+Branch type to add bounded flow variables and use flow constraints
+"""
+struct StaticBranchBounds <: AbstractBranchFormulation end
+"""
+Branch type to avoid flow constraints
+"""
+struct StaticBranchUnbounded <: AbstractBranchFormulation end
+
+############################### DC Branch Formulations #####################################
+abstract type AbstractDCLineFormulation <: AbstractBranchFormulation end
+struct HVDCUnbounded <: AbstractDCLineFormulation end
+struct HVDCLossless <: AbstractDCLineFormulation end
+struct HVDCDispatch <: AbstractDCLineFormulation end
+# Not Implemented
+# struct VoltageSourceDC <: AbstractDCLineFormulation end
+
+########################### Hybrid Generation Formulations ################################
+abstract type AbstractHybridFormulation <: AbstractDeviceFormulation end
+abstract type AbstractStandardHybridFormulation <: AbstractHybridFormulation end
+struct BasicHybridDispatch <: AbstractHybridFormulation end
+struct StandardHybridDispatch <: AbstractStandardHybridFormulation end
+
+############################ Hydro Generation Formulations #################################
+abstract type AbstractHydroFormulation <: AbstractDeviceFormulation end
+abstract type AbstractHydroDispatchFormulation <: AbstractHydroFormulation end
+abstract type AbstractHydroUnitCommitment <: AbstractHydroFormulation end
+abstract type AbstractHydroReservoirFormulation <: AbstractHydroDispatchFormulation end
+struct HydroDispatchRunOfRiver <: AbstractHydroDispatchFormulation end
+struct HydroDispatchReservoirBudget <: AbstractHydroReservoirFormulation end
+struct HydroDispatchReservoirStorage <: AbstractHydroReservoirFormulation end
+struct HydroDispatchPumpedStorage <: AbstractHydroReservoirFormulation end
+struct HydroCommitmentRunOfRiver <: AbstractHydroUnitCommitment end
+struct HydroCommitmentReservoirBudget <: AbstractHydroUnitCommitment end
+struct HydroCommitmentReservoirStorage <: AbstractHydroUnitCommitment end
+
+############################ Regulation Device Formulations ################################
+abstract type AbstractRegulationFormulation <: AbstractDeviceFormulation end
+struct ReserveLimitedRegulation <: AbstractRegulationFormulation end
+struct DeviceLimitedRegulation <: AbstractRegulationFormulation end
+
+########################### Renewable Generation Formulations ##############################
+abstract type AbstractRenewableFormulation <: AbstractDeviceFormulation end
+abstract type AbstractRenewableDispatchFormulation <: AbstractRenewableFormulation end
+struct RenewableFullDispatch <: AbstractRenewableDispatchFormulation end
+struct RenewableConstantPowerFactor <: AbstractRenewableDispatchFormulation end
+
+
+############################ Storage Generation Formulations ###############################
+abstract type AbstractStorageFormulation <: AbstractDeviceFormulation end
+abstract type AbstractEnergyManagement  <: AbstractStorageFormulation end
+struct BookKeeping <: AbstractStorageFormulation end
+struct BatteryAncillaryServices <: AbstractStorageFormulation end
+struct EnergyTarget <: AbstractEnergyManagement end
+
+
+############################## Network Model Formulations ##################################
+# These formulations are taken directly from PowerModels
+
+abstract type AbstractPTDFModel <: PM.AbstractDCPModel end
+struct StandardPTDFModel <: AbstractPTDFModel end
+struct PTDFPowerModel <: AbstractPTDFModel end
+
+struct CopperPlatePowerModel <: PM.AbstractActivePowerModel end
+struct AreaBalancePowerModel <: PM.AbstractActivePowerModel end
+
+#================================================
+    # exact non-convex models
+    ACPPowerModel, ACRPowerModel, ACTPowerModel
+
+    # linear approximations
+    DCPPowerModel, NFAPowerModel
+
+    # quadratic approximations
+    DCPLLPowerModel, LPACCPowerModel
+
+    # quadratic relaxations
+    SOCWRPowerModel, SOCWRConicPowerModel,
+    SOCBFPowerModel, SOCBFConicPowerModel,
+    QCRMPowerModel, QCLSPowerModel,
+
+    # sdp relaxations
+    SDPWRMPowerModel, SparseSDPWRMPowerModel
+================================================#
+
+##### Exact Non-Convex Models #####
+import PowerModels: ACPPowerModel
+
+import PowerModels: ACRPowerModel
+
+import PowerModels: ACTPowerModel
+
+##### Linear Approximations #####
+import PowerModels: DCPPowerModel
+
+import PowerModels: NFAPowerModel
+
+##### Quadratic Approximations #####
+import PowerModels: DCPLLPowerModel
+
+import PowerModels: LPACCPowerModel
+
+##### Quadratic Relaxations #####
+import PowerModels: SOCWRPowerModel
+
+import PowerModels: SOCWRConicPowerModel
+
+import PowerModels: QCRMPowerModel
+
+import PowerModels: QCLSPowerModel
+
+
+abstract type AbstractAGCFormulation <: AbstractServiceFormulation end
+struct PIDSmoothACE <: AbstractAGCFormulation end
+
+struct GroupReserve <: AbstractReservesFormulation end
+
+
+struct RangeReserve <: AbstractReservesFormulation end
+struct StepwiseCostReserve <: AbstractReservesFormulation end
+struct RampReserve <: AbstractReservesFormulation end
+struct NonSpinningReserve <: AbstractReservesFormulation end
